Split TimescaleListenStore
