--- listenbrainz/listenstore/tests/test_test.py
@@ -1,68 +1,0 @@
-# coding=utf-8
-import logging
-import os
-
-import psycopg2
-import sqlalchemy
-
-import listenbrainz.db.user as db_user
-from listenbrainz import config
-from listenbrainz.db import timescale as ts
-from listenbrainz.db.testing import DatabaseTestCase
-from listenbrainz.listenstore.tests.util import create_test_data_for_timescalelistenstore
-from listenbrainz.webserver import timescale_connection
-
-TIMESCALE_SQL_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', '..', 'admin', 'timescale')
-
-
-class TestTimescaleListenStore(DatabaseTestCase):
-
-    def reset_timescale_db(self):
-
-        ts.init_db_connection(config.TIMESCALE_ADMIN_URI)
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'drop_db.sql'))
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'create_db.sql'))
-        ts.engine.dispose()
-
-        ts.init_db_connection(config.TIMESCALE_ADMIN_LB_URI)
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'create_extensions.sql'))
-        ts.engine.dispose()
-
-        ts.init_db_connection(config.SQLALCHEMY_TIMESCALE_URI)
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_schemas.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_tables.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_functions.sql'))
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'create_views.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_indexes.sql'))
-        ts.create_view_indexes()
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_primary_keys.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_foreign_keys.sql'))
-        ts.engine.dispose()
-
-    def setUp(self):
-        super(TestTimescaleListenStore, self).setUp()
-        self.log = logging.getLogger(__name__)
-        self.reset_timescale_db()
-
-        self.ns = config.REDIS_NAMESPACE
-        self.logstore = timescale_connection._ts
-
-        self.testuser_id = db_user.create(1, "test")
-        self.testuser_name = db_user.get(self.testuser_id)['musicbrainz_id']
-
-    def tearDown(self):
-        self.logstore = None
-        super(TestTimescaleListenStore, self).tearDown()
-
-    def _create_test_data(self, user_name, test_data_file_name=None):
-        test_data = create_test_data_for_timescalelistenstore(user_name, 1, test_data_file_name)
-        self.logstore.insert(test_data)
-        return len(test_data)
-
-    def set_created_to_NULL_in_DB(self):
-        try:
-            with ts.engine.connect() as connection:
-                connection.execute(sqlalchemy.text("UPDATE listen SET created = NULL"))
-        except psycopg2.OperationalError as e:
-            self.log.error("Cannot query timescale listen_count: %s" % str(e), exc_info=True)
-            raise
