+++ listenbrainz/listenstore/tests/test_timescalelistenstore.py
@@ -1,64 +1,29 @@
-# coding=utf-8
-
 import logging
 import os
 import random
-import shutil
-import tarfile
-import tempfile
-from datetime import datetime
 from time import time
 
 import psycopg2
 import sqlalchemy
 from brainzutils import cache
-from psycopg2.extras import execute_values
 
 import listenbrainz.db.user as db_user
-from listenbrainz import config
 from listenbrainz.db import timescale as ts
-from listenbrainz.db.dump import SchemaMismatchException
-from listenbrainz.db.testing import DatabaseTestCase
-from listenbrainz.listenstore import LISTENS_DUMP_SCHEMA_VERSION
+from listenbrainz.db.testing import DatabaseTestCase, TimescaleTestCase
 from listenbrainz.listenstore.tests.util import create_test_data_for_timescalelistenstore, generate_data
-from listenbrainz.listenstore.timescale_listenstore import REDIS_USER_LISTEN_COUNT, REDIS_USER_TIMESTAMPS
-from listenbrainz.webserver import timescale_connection
+from listenbrainz.listenstore.timescale_listenstore import REDIS_USER_LISTEN_COUNT, REDIS_USER_TIMESTAMPS, \
+    TimescaleListenStore
 
 TIMESCALE_SQL_DIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', '..', 'admin', 'timescale')
 
 
-class TestTimescaleListenStore(DatabaseTestCase):
-
-    def reset_timescale_db(self):
-
-        ts.init_db_connection(config.TIMESCALE_ADMIN_URI)
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'drop_db.sql'))
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'create_db.sql'))
-        ts.engine.dispose()
-
-        ts.init_db_connection(config.TIMESCALE_ADMIN_LB_URI)
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'create_extensions.sql'))
-        ts.engine.dispose()
-
-        ts.init_db_connection(config.SQLALCHEMY_TIMESCALE_URI)
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_schemas.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_types.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_tables.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_functions.sql'))
-        ts.run_sql_script_without_transaction(os.path.join(TIMESCALE_SQL_DIR, 'create_views.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_indexes.sql'))
-        ts.create_view_indexes()
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_primary_keys.sql'))
-        ts.run_sql_script(os.path.join(TIMESCALE_SQL_DIR, 'create_foreign_keys.sql'))
-        ts.engine.dispose()
+class TestTimescaleListenStore(DatabaseTestCase, TimescaleTestCase):
 
     def setUp(self):
-        super(TestTimescaleListenStore, self).setUp()
+        DatabaseTestCase.setUp(self)
+        TimescaleTestCase.setUp(self)
         self.log = logging.getLogger(__name__)
-        self.reset_timescale_db()
-
-        self.ns = config.REDIS_NAMESPACE
-        self.logstore = timescale_connection._ts
+        self.logstore = TimescaleListenStore(self.log)
 
         self.testuser = db_user.get_or_create(1, "test")
         self.testuser_id = self.testuser["id"]
@@ -66,8 +31,9 @@ class TestTimescaleListenStore(DatabaseTestCase):
 
     def tearDown(self):
         self.logstore = None
+        DatabaseTestCase.tearDown(self)
+        TimescaleTestCase.tearDown(self)
         cache._r.flushdb()
-        super(TestTimescaleListenStore, self).tearDown()
 
     def _create_test_data(self, user_name, user_id, test_data_file_name=None):
         test_data = create_test_data_for_timescalelistenstore(user_name, user_id, test_data_file_name)
@@ -94,25 +60,6 @@ class TestTimescaleListenStore(DatabaseTestCase):
         with ts.engine.connect() as connection:
             connection.execute(sqlalchemy.text(query))
             connection.execute(sqlalchemy.text(join_query))
-
-    def _insert_with_created(self, listens):
-        """ Insert a batch of listens with 'created' field.
-        """
-        submit = []
-        for listen in listens:
-            submit.append((*listen.to_timescale(), listen.inserted_timestamp))
-
-        query = """INSERT INTO listen (listened_at, track_name, user_name, user_id, data, created)
-                        VALUES %s
-                   ON CONFLICT (listened_at, track_name, user_id)
-                    DO NOTHING
-                """
-
-        conn = ts.engine.raw_connection()
-        with conn.cursor() as curs:
-            execute_values(curs, query, submit, template=None)
-
-        conn.commit()
 
     def test_check_listen_count_view_exists(self):
         try:
@@ -247,181 +194,6 @@ class TestTimescaleListenStore(DatabaseTestCase):
                                                               recent[0].ts_since_epoch + 1)
         self.assertEqual(len(recent), 1)
         self.assertEqual(recent[0].ts_since_epoch, 1400000200)
-
-    def test_dump_listens(self):
-        self._create_test_data(self.testuser_name, self.testuser_id)
-        temp_dir = tempfile.mkdtemp()
-        dump = self.logstore.dump_listens(
-            location=temp_dir,
-            dump_id=1,
-            end_time=datetime.now(),
-        )
-        self.assertTrue(os.path.isfile(dump))
-        shutil.rmtree(temp_dir)
-
-    def test_incremental_dump(self):
-        base = 1500000000
-        listens = generate_data(1, self.testuser_name, base-4, 5, base+1)  # generate 5 listens with inserted_ts 1-5
-        self._insert_with_created(listens)
-        listens = generate_data(1, self.testuser_name, base+1, 5, base+6)  # generate 5 listens with inserted_ts 6-10
-        self._insert_with_created(listens)
-        temp_dir = tempfile.mkdtemp()
-        dump_location = self.logstore.dump_listens(
-            location=temp_dir,
-            dump_id=1,
-            start_time=datetime.utcfromtimestamp(base + 6),
-            end_time=datetime.utcfromtimestamp(base + 10)
-        )
-        self.assertTrue(os.path.isfile(dump_location))
-        self.reset_timescale_db()
-        self.logstore.import_listens_dump(dump_location)
-        listens, min_ts, max_ts = self.logstore.fetch_listens(user=self.testuser, to_ts=base + 11)
-        self.assertEqual(len(listens), 4)
-        self.assertEqual(listens[0].ts_since_epoch, base + 5)
-        self.assertEqual(listens[1].ts_since_epoch, base + 4)
-        self.assertEqual(listens[2].ts_since_epoch, base + 3)
-        self.assertEqual(listens[3].ts_since_epoch, base + 2)
-
-        shutil.rmtree(temp_dir)
-
-    def test_time_range_full_dumps(self):
-        base = 1500000000
-        listens = generate_data(1, self.testuser_name, base + 1, 5)  # generate 5 listens with ts 1-5
-        self.logstore.insert(listens)
-        listens = generate_data(1, self.testuser_name, base + 6, 5)  # generate 5 listens with ts 6-10
-        self.logstore.insert(listens)
-        temp_dir = tempfile.mkdtemp()
-        dump_location = self.logstore.dump_listens(
-            location=temp_dir,
-            dump_id=1,
-            end_time=datetime.utcfromtimestamp(base + 5)
-        )
-        self.assertTrue(os.path.isfile(dump_location))
-        self.reset_timescale_db()
-        self.logstore.import_listens_dump(dump_location)
-        listens, min_ts, max_ts = self.logstore.fetch_listens(user=self.testuser, to_ts=base + 11)
-        self.assertEqual(len(listens), 5)
-        self.assertEqual(listens[0].ts_since_epoch, base + 5)
-        self.assertEqual(listens[1].ts_since_epoch, base + 4)
-        self.assertEqual(listens[2].ts_since_epoch, base + 3)
-        self.assertEqual(listens[3].ts_since_epoch, base + 2)
-        self.assertEqual(listens[4].ts_since_epoch, base + 1)
-
-    # tests test_full_dump_listen_with_no_created
-    # and test_incremental_dumps_listen_with_no_created have been removed because
-    # with timescale all the missing inserted timestamps will have been
-    # been assigned sane created timestamps by the migration script
-    # and timescale will not allow blank created timestamps, so this test is pointless
-
-    def test_import_listens(self):
-        self._create_test_data(self.testuser_name, self.testuser_id)
-        temp_dir = tempfile.mkdtemp()
-        dump_location = self.logstore.dump_listens(
-            location=temp_dir,
-            dump_id=1,
-            end_time=datetime.now(),
-        )
-        self.assertTrue(os.path.isfile(dump_location))
-        self.reset_timescale_db()
-        self.logstore.import_listens_dump(dump_location)
-        listens, min_ts, max_ts = self.logstore.fetch_listens(user=self.testuser, to_ts=1400000300)
-        self.assertEqual(len(listens), 5)
-        self.assertEqual(listens[0].ts_since_epoch, 1400000200)
-        self.assertEqual(listens[1].ts_since_epoch, 1400000150)
-        self.assertEqual(listens[2].ts_since_epoch, 1400000100)
-        self.assertEqual(listens[3].ts_since_epoch, 1400000050)
-        self.assertEqual(listens[4].ts_since_epoch, 1400000000)
-        shutil.rmtree(temp_dir)
-
-    def test_dump_and_import_listens_escaped(self):
-        user = db_user.get_or_create(3, 'i have a\\weird\\user, na/me"\n')
-        self._create_test_data(user['musicbrainz_id'], user['id'])
-
-        self._create_test_data(self.testuser_name, self.testuser_id)
-
-        temp_dir = tempfile.mkdtemp()
-        dump_location = self.logstore.dump_listens(
-            location=temp_dir,
-            dump_id=1,
-            end_time=datetime.now(),
-        )
-        self.assertTrue(os.path.isfile(dump_location))
-        self.reset_timescale_db()
-
-        self.logstore.import_listens_dump(dump_location)
-
-        listens, min_ts, max_ts = self.logstore.fetch_listens(user=user, to_ts=1400000300)
-        self.assertEqual(len(listens), 5)
-        self.assertEqual(listens[0].ts_since_epoch, 1400000200)
-        self.assertEqual(listens[1].ts_since_epoch, 1400000150)
-        self.assertEqual(listens[2].ts_since_epoch, 1400000100)
-        self.assertEqual(listens[3].ts_since_epoch, 1400000050)
-        self.assertEqual(listens[4].ts_since_epoch, 1400000000)
-
-        listens, min_ts, max_ts = self.logstore.fetch_listens(user=self.testuser, to_ts=1400000300)
-        self.assertEqual(len(listens), 5)
-        self.assertEqual(listens[0].ts_since_epoch, 1400000200)
-        self.assertEqual(listens[1].ts_since_epoch, 1400000150)
-        self.assertEqual(listens[2].ts_since_epoch, 1400000100)
-        self.assertEqual(listens[3].ts_since_epoch, 1400000050)
-        self.assertEqual(listens[4].ts_since_epoch, 1400000000)
-        shutil.rmtree(temp_dir)
-
-    # test test_import_dump_many_users is gone -- why are we testing user dump/restore here??
-
-    def create_test_dump(self, archive_name, archive_path, schema_version=None):
-        """ Creates a test dump to test the import listens functionality.
-
-        Args:
-            archive_name (str): the name of the archive
-            archive_path (str): the full path to the archive
-            schema_version (int): the version of the schema to be written into SCHEMA_SEQUENCE
-                                  if not provided, the SCHEMA_SEQUENCE file is not added to the archive
-
-        Returns:
-            the full path to the archive created
-        """
-
-        temp_dir = tempfile.mkdtemp()
-        with tarfile.open(archive_path, mode='w|xz') as tar:
-            schema_version_path = os.path.join(temp_dir, 'SCHEMA_SEQUENCE')
-            with open(schema_version_path, 'w') as f:
-                f.write(str(schema_version or ' '))
-            tar.add(schema_version_path,
-                    arcname=os.path.join(archive_name, 'SCHEMA_SEQUENCE'))
-
-        return archive_path
-
-    def test_schema_mismatch_exception_for_dump_incorrect_schema(self):
-        """ Tests that SchemaMismatchException is raised when the schema of the dump is old """
-
-        # create a temp archive with incorrect SCHEMA_VERSION_CORE
-        temp_dir = tempfile.mkdtemp()
-        archive_name = 'temp_dump'
-        archive_path = os.path.join(temp_dir, archive_name + '.tar.xz')
-        archive_path = self.create_test_dump(
-            archive_name=archive_name,
-            archive_path=archive_path,
-            schema_version=LISTENS_DUMP_SCHEMA_VERSION - 1
-        )
-        with self.assertRaises(SchemaMismatchException):
-            self.logstore.import_listens_dump(archive_path)
-
-    def test_schema_mismatch_exception_for_dump_no_schema(self):
-        """ Tests that SchemaMismatchException is raised when there is no schema version in the archive """
-
-        temp_dir = tempfile.mkdtemp()
-        archive_name = 'temp_dump'
-        archive_path = os.path.join(temp_dir, archive_name + '.tar.xz')
-
-        archive_path = self.create_test_dump(
-            archive_name=archive_name,
-            archive_path=archive_path,
-            schema_version=None
-        )
-
-        with self.assertRaises(SchemaMismatchException):
-            self.logstore.import_listens_dump(archive_path)
 
     def test_listen_counts_in_cache(self):
         uid = random.randint(2000, 1 << 31)
